backend/requirements.txt
Django==5.2.6
djangorestframework==3.16.1
django-cors-headers==4.7.0
mssql-django==1.7
pyodbc==5.2.0
